refactor: handlers to smaller modules docs: update documentation in modules
