+++ app/lib/router.js
@@ -1,19 +1,13 @@
 /**
- * Router for all requests
+ * @module Router
+ * @desc Router for all requests
  */
 
 const handlers = require("./handlers");
 
-/**
- *
- * @type {Record<string, (function(data: any, callback: function(statusCode:
- *     string, payload?: any)): void)>}
- */
-const router = {
+module.exports = {
     ping: handlers.ping,
     checks: handlers.checks,
     tokens: handlers.tokens,
     users: handlers.users,
 };
-module.exports = router;
-
